Module Jsonrpc2_core.Protocol

The protocol part, independent from IO and Transport

type json = J.t

type t

A jsonrpc2 connection.

val create : unit -> t

Create a state machine for Jsonrpc2

val clear : t -> unit

Clear all internal state.

module Id : sig ... end

Send requests and notifications to the other side

type message = json

Message sent to the other side

val error : t -> code -> string -> message

val request : t -> meth:string -> params:json option -> message * Id.t

Create a request message, for which an answer is expected.

val notify : t -> meth:string -> params:json option -> message

Create a notification message, ie. no response is expected.

type action =

| Send of message

| Send_batch of message list

| Start_call of Id.t * string * json option

| Notify of string * json option

| Fill_request of Id.t * (json, int * string) Pervasives.result

| Error_without_id of int * string

Actions to be done next. This includes sending messages out on the connection, calling a method, or finishing a local request.

val process_msg : t -> message -> (action list, code * string) Pervasives.result

Process incoming message

val process_call_reply : t -> Id.t -> (json, string) Pervasives.result -> action list

Send the response for the given call to the other side