Module type Jsonrpc2__Jsonrpc2_intf.S

Signature for a JSONRPC-2 implementation

module IO : IO

type json = J.t

type t

A jsonrpc2 connection.

val create : ic:IO.in_channel -> oc:IO.out_channel -> unit -> t

Create a connection from the pair of channels

Declare methods available from the other side

type method_ = t -> params:json option -> return:((json, string) Pervasives.result -> unit) -> unit

val declare_method : t -> string -> method_ -> unit

Add a method that can be called from the other side. The method, when called, must at some point call its return paramter with a result.

val declare_method_with : t -> decode_arg:(json -> ('a, string) Pervasives.result) -> encode_res:('b -> json) -> string -> (t -> params:'a option -> return:('b -> unit) -> unit) -> unit

Sugar around declare_method, with automatic encoding and decoding of JSON values.

val declare_blocking_method_with : t -> decode_arg:(json -> ('a, string) Pervasives.result) -> encode_res:('b -> json) -> string -> ('a option -> 'b) -> unit

Sugar around declare_method_with when the function returns quickly (no scheduling in the background), with automatic encoding and decoding of JSON values. The function is invoked immediately and returns a value that is then fed to the underlying return callback.

Send requests and notifications to the other side

exception Jsonrpc2_error of int * string

Code + message

type message

Message sent to the other side

val request : t -> meth:string -> params:json option -> message * (json, int * string) Pervasives.result IO.Future.t

Create a request message, for which an answer is expected.

val notify : t -> meth:string -> params:json option -> message

Create a notification message, ie. no response is expected.

val send : t -> message -> (unit, exn) Pervasives.result IO.t

Send the message.

val send_batch : t -> message list -> (unit, exn) Pervasives.result IO.t

Send a batch of messages.

val send_request : t -> meth:string -> params:json option -> (json, exn) Pervasives.result IO.t

Combination of send and request

val send_notify : t -> meth:string -> params:json option -> (unit, exn) Pervasives.result IO.t

Combination of send and notify

val send_request_with : encode_params:('a -> json) -> decode_res:(json -> ('b, string) Pervasives.result) -> t -> meth:string -> params:'a option -> ('b, exn) Pervasives.result IO.t

Decoders + send_request

val send_notify_with : encode_params:('a -> json) -> t -> meth:string -> params:'a option -> (unit, exn) Pervasives.result IO.t

Encoder + send_notify

val run : t -> (unit, exn) Pervasives.result IO.t

Listen for incoming messages and responses